# docs/swagger.yaml
definitions:
...
  collegemanagementsystem.Credentials:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  collegemanagementsystem.Lecturer:
    properties:
      age:
        type: integer
      designation:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/lecturers:
    get:
      description: Retrieve all lecturers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/collegemanagementsystem.Lecturer'
            type: array
      security:
      - BearerAuth: []
      summary: Get all lecturers
      tags:
      - Lecturers
    post:
      consumes:
      - application/json
      parameters:
      - description: Lecturer Data
        in: body
        name: lecturer
        required: true
        schema:
          $ref: '#/definitions/collegemanagementsystem.Lecturer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/collegemanagementsystem.Lecturer'
      security:
      - BearerAuth: []
      summary: Create lecturer
      tags:
      - Lecturers
  /api/lecturers/{id}:
    delete:
      parameters:
      - description: Lecturer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: lecturer deleted
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete lecturer
      tags:
      - Lecturers
    get:
      parameters:
      - description: Lecturer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/collegemanagementsystem.Lecturer'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get lecturer by ID
      tags:
      - Lecturers
    put:
      consumes:
      - application/json
      parameters:
      - description: Updated Lecturer
        in: body
        name: lecturer
        required: true
        schema:
          $ref: '#/definitions/collegemanagementsystem.Lecturer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/collegemanagementsystem.Lecturer'
      security:
      - BearerAuth: []
      summary: Update lecturer
      tags:
      - Lecturers
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens in cookies
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/collegemanagementsystem.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Authentication
...
  /refresh:
    post:
      description: Generate new access token using refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - Authentication
swagger: "2.0"